combine cdata_testframework.c into cdata_fulltest.c, rename to is_little_endian
